fix circular include
